+++ src/ipc/test/ipc_tests.cpp
@@ -5,6 +5,7 @@
 #include <ipc/process.h>
 #include <ipc/test/ipc_test.h>
 
+#include <test/util/common.h>
 #include <test/util/setup_common.h>
 #include <boost/test/unit_test.hpp>
