+++ src/test/util/common.h
@@ -9,6 +9,22 @@
 #include <optional>
 #include <string>
 
+/**
+ * BOOST_CHECK_EXCEPTION predicates to check the specific validation error.
+ * Use as
+ * BOOST_CHECK_EXCEPTION(code that throws, exception type, HasReason("foo"));
+ */
+class HasReason
+{
+public:
+    explicit HasReason(std::string_view reason) : m_reason(reason) {}
+    bool operator()(std::string_view s) const { return s.find(m_reason) != std::string_view::npos; }
+    bool operator()(const std::exception& e) const { return (*this)(e.what()); }
+
+private:
+    const std::string m_reason;
+};
+
 // Make types usable in BOOST_CHECK_* @{
 namespace std {
 template <typename T> requires std::is_enum_v<T>
