+++ src/test/util_check_tests.cpp
@@ -5,7 +5,7 @@
 #include <util/check.h>
 
 #include <boost/test/unit_test.hpp>
-#include <test/util/setup_common.h>
+#include <test/util/common.h>
 
 BOOST_AUTO_TEST_SUITE(util_check_tests)
